Move caching to cache.py and Mode/TargetVersion/Feature to mode.py
