+++ src/black/__init__.py
@@ -13,11 +13,9 @@ import logging
 from multiprocessing import Manager, freeze_support
 import os
 from pathlib import Path
-import pickle
 import regex as re
 import signal
 import sys
-import tempfile
 import tokenize
 import traceback
 from typing import (
@@ -48,7 +46,7 @@ import click
 import toml
 
 from black.const import DEFAULT_LINE_LENGTH, DEFAULT_INCLUDES, DEFAULT_EXCLUDES
-from black.const import CACHE_DIR, STDIN_PLACEHOLDER, STRING_PREFIX_CHARS
+from black.const import STDIN_PLACEHOLDER, STRING_PREFIX_CHARS
 from black.nodes import WHITESPACE, STATEMENT, STANDALONE_COMMENT, LOGIC_OPERATORS
 from black.nodes import COMPARATORS, MATH_OPERATORS, STARS, VARARGS_PARENTS
 from black.nodes import UNPACKING_PARENTS, TEST_DESCENDANTS, ASSIGNMENTS
@@ -65,6 +63,9 @@ from black.strings import format_hex, format_scientific_notation
 from black.strings import format_long_or_complex_number, format_float_or_int_string
 from black.strings import dump_to_file, diff, color_diff
 from black.strings import has_triple_quotes, fix_docstring
+from black.mode import Mode, TargetVersion
+from black.mode import Feature, supports_feature, VERSION_TO_FEATURES
+from black.cache import read_cache, write_cache, get_cache_info, filter_cached, Cache
 
 try:
     from typed_ast import ast3, ast27
@@ -112,10 +113,6 @@ Priority = int
 Index = int
 LN = Union[Leaf, Node]
 Transformer = Callable[["Line", Collection["Feature"]], Iterator["Line"]]
-Timestamp = float
-FileSize = int
-CacheInfo = Tuple[Timestamp, FileSize]
-Cache = Dict[str, CacheInfo]
 out = partial(click.secho, bold=True, err=True)
 err = partial(click.secho, fg="red", err=True)
 
@@ -194,120 +191,8 @@ class Changed(Enum):
     YES = 2
 
 
-class TargetVersion(Enum):
-    PY27 = 2
-    PY33 = 3
-    PY34 = 4
-    PY35 = 5
-    PY36 = 6
-    PY37 = 7
-    PY38 = 8
-    PY39 = 9
-
-    def is_python2(self) -> bool:
-        return self is TargetVersion.PY27
-
-
-class Feature(Enum):
-    # All string literals are unicode
-    UNICODE_LITERALS = 1
-    F_STRINGS = 2
-    NUMERIC_UNDERSCORES = 3
-    TRAILING_COMMA_IN_CALL = 4
-    TRAILING_COMMA_IN_DEF = 5
-    # The following two feature-flags are mutually exclusive, and exactly one should be
-    # set for every version of python.
-    ASYNC_IDENTIFIERS = 6
-    ASYNC_KEYWORDS = 7
-    ASSIGNMENT_EXPRESSIONS = 8
-    POS_ONLY_ARGUMENTS = 9
-    RELAXED_DECORATORS = 10
-    FORCE_OPTIONAL_PARENTHESES = 50
-
-
-VERSION_TO_FEATURES: Dict[TargetVersion, Set[Feature]] = {
-    TargetVersion.PY27: {Feature.ASYNC_IDENTIFIERS},
-    TargetVersion.PY33: {Feature.UNICODE_LITERALS, Feature.ASYNC_IDENTIFIERS},
-    TargetVersion.PY34: {Feature.UNICODE_LITERALS, Feature.ASYNC_IDENTIFIERS},
-    TargetVersion.PY35: {
-        Feature.UNICODE_LITERALS,
-        Feature.TRAILING_COMMA_IN_CALL,
-        Feature.ASYNC_IDENTIFIERS,
-    },
-    TargetVersion.PY36: {
-        Feature.UNICODE_LITERALS,
-        Feature.F_STRINGS,
-        Feature.NUMERIC_UNDERSCORES,
-        Feature.TRAILING_COMMA_IN_CALL,
-        Feature.TRAILING_COMMA_IN_DEF,
-        Feature.ASYNC_IDENTIFIERS,
-    },
-    TargetVersion.PY37: {
-        Feature.UNICODE_LITERALS,
-        Feature.F_STRINGS,
-        Feature.NUMERIC_UNDERSCORES,
-        Feature.TRAILING_COMMA_IN_CALL,
-        Feature.TRAILING_COMMA_IN_DEF,
-        Feature.ASYNC_KEYWORDS,
-    },
-    TargetVersion.PY38: {
-        Feature.UNICODE_LITERALS,
-        Feature.F_STRINGS,
-        Feature.NUMERIC_UNDERSCORES,
-        Feature.TRAILING_COMMA_IN_CALL,
-        Feature.TRAILING_COMMA_IN_DEF,
-        Feature.ASYNC_KEYWORDS,
-        Feature.ASSIGNMENT_EXPRESSIONS,
-        Feature.POS_ONLY_ARGUMENTS,
-    },
-    TargetVersion.PY39: {
-        Feature.UNICODE_LITERALS,
-        Feature.F_STRINGS,
-        Feature.NUMERIC_UNDERSCORES,
-        Feature.TRAILING_COMMA_IN_CALL,
-        Feature.TRAILING_COMMA_IN_DEF,
-        Feature.ASYNC_KEYWORDS,
-        Feature.ASSIGNMENT_EXPRESSIONS,
-        Feature.RELAXED_DECORATORS,
-        Feature.POS_ONLY_ARGUMENTS,
-    },
-}
-
-
-@dataclass
-class Mode:
-    target_versions: Set[TargetVersion] = field(default_factory=set)
-    line_length: int = DEFAULT_LINE_LENGTH
-    string_normalization: bool = True
-    is_pyi: bool = False
-    magic_trailing_comma: bool = True
-    experimental_string_processing: bool = False
-
-    def get_cache_key(self) -> str:
-        if self.target_versions:
-            version_str = ",".join(
-                str(version.value)
-                for version in sorted(self.target_versions, key=lambda v: v.value)
-            )
-        else:
-            version_str = "-"
-        parts = [
-            version_str,
-            str(self.line_length),
-            str(int(self.string_normalization)),
-            str(int(self.is_pyi)),
-            str(int(self.magic_trailing_comma)),
-            str(int(self.experimental_string_processing)),
-        ]
-        return ".".join(parts)
-
-
 # Legacy name, left for integrations.
 FileMode = Mode
-
-
-def supports_feature(target_versions: Set[TargetVersion], feature: Feature) -> bool:
-    return all(feature in VERSION_TO_FEATURES[version] for version in target_versions)
 
 
 def find_pyproject_toml(path_search_start: Tuple[str, ...]) -> Optional[str]:
@@ -5875,66 +5760,6 @@ def run_transformer(
     ):
         result = second_opinion
     return result
-
-
-def get_cache_file(mode: Mode) -> Path:
-    return CACHE_DIR / f"cache.{mode.get_cache_key()}.pickle"
-
-
-def read_cache(mode: Mode) -> Cache:
-    """Read the cache if it exists and is well formed.
-
-    If it is not well formed, the call to write_cache later should resolve the issue.
-    """
-    cache_file = get_cache_file(mode)
-    if not cache_file.exists():
-        return {}
-
-    with cache_file.open("rb") as fobj:
-        try:
-            cache: Cache = pickle.load(fobj)
-        except (pickle.UnpicklingError, ValueError):
-            return {}
-
-    return cache
-
-
-def get_cache_info(path: Path) -> CacheInfo:
-    """Return the information used to check if a file is already formatted or not."""
-    stat = path.stat()
-    return stat.st_mtime, stat.st_size
-
-
-def filter_cached(cache: Cache, sources: Iterable[Path]) -> Tuple[Set[Path], Set[Path]]:
-    """Split an iterable of paths in `sources` into two sets.
-
-    The first contains paths of files that modified on disk or are not in the
-    cache. The other contains paths to non-modified files.
-    """
-    todo, done = set(), set()
-    for src in sources:
-        res_src = src.resolve()
-        if cache.get(str(res_src)) != get_cache_info(res_src):
-            todo.add(src)
-        else:
-            done.add(src)
-    return todo, done
-
-
-def write_cache(cache: Cache, sources: Iterable[Path], mode: Mode) -> None:
-    """Update the cache file."""
-    cache_file = get_cache_file(mode)
-    try:
-        CACHE_DIR.mkdir(parents=True, exist_ok=True)
-        new_cache = {
-            **cache,
-            **{str(src.resolve()): get_cache_info(src) for src in sources},
-        }
-        with tempfile.NamedTemporaryFile(dir=str(cache_file.parent), delete=False) as f:
-            pickle.dump(new_cache, f, protocol=4)
-        os.replace(f.name, cache_file)
-    except OSError:
-        pass
 
 
 def patch_click() -> None:
